Improved and Added Tests Improved LaTeX special symbols escaping Added support for: - Strikethrough - Superscript - Subscript - Emphasis - Hyperlinks - h4,h5,h6 - Unrecongised tags
