+++ package.json
@@ -23,7 +23,7 @@
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-jest": "^23.8.2",
     "eslint-plugin-prettier": "^3.1.3",
-    "jest": "^25.4.0",
+    "jest": "^29.7.0",
     "prettier": "^2.0.5",
     "rollup": "^2.7.3",
     "rollup-plugin-preserve-shebangs": "^0.2.0",
